backend/requirements.txt
Django==3.1.6
PyMySQL==1.0.2
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0